requirements.txt
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
exceptiongroup==1.2.0
fastapi==0.104.1
fonttools==4.45.1
future==0.18.3
h11==0.14.0
idna==3.6
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.2
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.1.0
pydantic==2.5.2
pydantic_core==2.14.5
pyparsing==3.1.1
python-dateutil==2.8.2
python-multipart==0.0.6
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tk==0.1.0
tkintertable==1.3.3
typing_extensions==4.8.0
uvicorn==0.24.0.post1
webcolors==1.13
